requirements.txt
# #testing
# pytest==8.0.1

# # for server
# Hypercorn==0.16.0
# fastapi==0.109.0
# asyncio==3.4.3
# python-multipart==0.0.6

# for job procs 
pandas==2.1.4
requests==2.31.0
lxml==5.1.0
python-dotenv==1.0.0

# # playwright==1.41.2
# beautifulsoup4==4.12.3
# langchain==0.0.352
# openai==1.10.0
# langchain-openai==0.0.5
# pydantic==2.5.3
# pydantic_core==2.14.6
# python-docx==1.1.0
# htmldocx==0.0.6
# google-cloud-storage==2.14.0
# google-auth==2.27.0
# google-api-python-client==2.116.0
# google-cloud-bigquery
# google-search-results==2.4.2 ## for serp api
# langchain-pinecone==0.0.2

# # google ai's
# langchain-google-genai==0.0.9
# langchain-google-vertexai==0.0.5